requirements.txt
amqp==2.6.1
asgiref==3.3.1
Babel==2.9.0
billiard==3.6.3.0
braintree==3.59.0
cairocffi==1.2.0
CairoSVG==2.5.0
celery==4.4.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
cssselect2==0.4.1
defusedxml==0.7.0rc1
Django==3.1.3
django-crispy-forms==1.10.0
django-extensions==3.0.9
django-localflavor==3.0.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
flower==0.9.3
gunicorn==20.0.4
html5lib==1.1
idna==2.10
kombu==4.6.11
localflavor==1.8
oauthlib==3.1.0
phonenumberslite==8.12.13
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
Pyphen==0.10.0
python-dotenv==0.15.0
python-stdnum==1.14
python3-openid==3.2.0
pytz==2020.4
pyzipcode==2.0.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.11.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
tinycss2==1.1.0
tornado==5.1.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==52.1
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
